checking if the scoreboard is deleted before updating the lines

+++ nucleo-modules/module-scoreboard/scoreboard-spigot/src/main/java/de/keeeks/nucleo/modules/scoreboard/spigot/NucleoScoreboard.java
@@ -104,6 +104,7 @@ public class NucleoScoreboard implements Scoreboard {
     @Override
     public void destroy() {
         player.removeMetadata("scoreboardId", module.plugin());
+        if (fastBoard.isDeleted()) return;
         fastBoard.delete();
     }
